Add some additional work around frame size limits and handling
Track frame size limits using longs to capture the value as non-negative
but enforce that the engine won't allow for a larger than 2GB outbound
max frame size since we cannot allocate buffers larger than that anyway.

+++ protonj2/src/main/java/org/apache/qpid/protonj2/engine/impl/ProtonFrameEncodingHandler.java
@@ -72,7 +72,7 @@ public class ProtonFrameEncodingHandler implements EngineHandler {
 
     @Override
     public void handleWrite(EngineHandlerContext context, SaslFrame frame) {
-        ProtonBuffer output = configuration.getBufferAllocator().outputBuffer(AMQP_PERFORMATIVE_PAD, configuration.getOutboundMaxFrameSize());
+        ProtonBuffer output = configuration.getBufferAllocator().outputBuffer(AMQP_PERFORMATIVE_PAD, (int) configuration.getOutboundMaxFrameSize());
 
         output.setWriteIndex(FRAME_HEADER_SIZE);
         output.setBytes(FRAME_START_BYTE, SASL_FRAME_HEADER);
@@ -91,10 +91,8 @@ public class ProtonFrameEncodingHandler implements EngineHandler {
     @Override
     public void handleWrite(EngineHandlerContext context, OutgoingProtocolFrame frame) {
         try {
-            // TODO: Ensure no numeric overflows in size calculations
-
             final ProtonBuffer payload = frame.getPayload() == null ? EMPTY_BUFFER : frame.getPayload();
-            final int maxFrameSize = configuration.getOutboundMaxFrameSize();
+            final int maxFrameSize = (int) configuration.getOutboundMaxFrameSize();
             final int outputBufferSize = Math.min(maxFrameSize, AMQP_PERFORMATIVE_PAD + payload.getReadableBytes());
             final ProtonBuffer output = configuration.getBufferAllocator().outputBuffer(outputBufferSize, maxFrameSize);
